created 3DOF hybrid robot class
