+++ dynamic/Hybrid_Manipulator.py
@@ -124,7 +124,7 @@ class HybridOneLinkManipulator( M.OneLinkManipulator ) :
         
         
       
- #########################################################################################################   
+#########################################################################################################   
 #########################################################################################################  
         
 
@@ -178,10 +178,10 @@ class HybridTwoLinkManipulator( M.TwoLinkManipulator ) :
         
         
     ##############################
-    def T(self, q = np.zeros(2) , dq = np.zeros(2) , ddq = np.zeros(2) , u3 = 0 ):
+    def T(self, q = np.zeros(2) , dq = np.zeros(2) , ddq = np.zeros(2) , uD = 0 ):
         """ Computed acutator torques given a trajectory and gear ratio """ 
         
-        R = self.R[u3]
+        R = self.R[uD]
         
         F = self.F( q , dq , ddq )
         
@@ -193,10 +193,10 @@ class HybridTwoLinkManipulator( M.TwoLinkManipulator ) :
         
         
     ##############################
-    def ddq_a(self, q = np.zeros(2) , dq = np.zeros(2) , T = np.zeros(2) , u3 = 0 ):
+    def ddq_a(self, q = np.zeros(2) , dq = np.zeros(2) , T = np.zeros(2) , uD = 0 ):
         """ Computed accelerations given actuator torques and gear ratio """  
         
-        R = self.R[ int(u3) ]
+        R = self.R[ int(uD) ]
         
         Ha = self.H( q ) + np.dot( R , np.dot( R , self.Ia ) )
         
@@ -230,26 +230,170 @@ class HybridTwoLinkManipulator( M.TwoLinkManipulator ) :
         
         dx = np.zeros(self.n) # State derivative vector
         
-        q  = x[0:2]
-        dq = x[2:4]
+        [ q , dq ] = self.x2q( x )   # from state vector (x) to angle and speeds (q,dq)
         
-        ddq = self.ddq_a( q , dq , u[0:2] , u[2] )
+        ddq = self.ddq_a( q , dq , u[0:self.dof] , u[self.dof] )
         
-        dx[0:2] = dq
-        dx[2:4] = ddq
+        dx = self.q2x( dq , ddq )    # from angle and speeds diff (dq,ddq) to state vector diff (dx)
         
         return dx
         
         
     ##############################
-    def e_kinetic(self, q = np.zeros(2) , dq = np.zeros(2) , u3 = 0 ):
+    def e_kinetic(self, q = np.zeros(2) , dq = np.zeros(2) , uD = 0 ):
         """ Compute kinetic energy of manipulator """  
         
-        R = self.R[ int(u3) ]
+        R = self.R[ int(uD) ]
         
         Ha = self.H( q ) + np.dot( R , np.dot( R , self.Ia ) )        
         
         e_k = 0.5 * np.dot( dq , np.dot( Ha , dq ) )
         
         return e_k
+        
+        
+        
+#########################################################################################################   
+#########################################################################################################  
+        
+
+
+class HybridThreeLinkManipulator( M.ThreeLinkManipulator ) :
+    """ 3DOF Manipulator Class """
+    
+    
+    ############################
+    def __init__(self, n = 6 , m = 4 ):
+        
+        M.ThreeLinkManipulator.__init__(self, n , m-1 )
+        
+        self.n = n
+        self.m = m        
+        
+        self.input_label = ['Torque 1','Torque 2','Torque 3','Mode']
+        self.input_units = ['[Nm]','[Nm]','[Nm]','']
+        
+        tmax = 5
+        
+        self.u_ub = np.array([ tmax, tmax, tmax, 3])      # Control Upper Bounds
+        self.u_lb = np.array([-tmax,-tmax,-tmax, 0])      # Control Lower Bounds
+        
+        self.ubar = np.array([0,0,0,0])
+        
+        self.setActuatorParams()
+        
+
+    #############################
+    def setActuatorParams(self, r1 = 1 , r2 = 10 , ja = 1 , ba = 1 ):
+        """ Set model parameters here """
+        
+        # Gear ratio
+        self.R = [np.diag([r1,r1,r1]),np.diag([r1,r2,r1]),np.diag([r2,r1,r1]),np.diag([r2,r2,r1])]
+        
+        # Inertia
+        self.Ia = np.diag([ja,ja,ja])
+        
+        # Damping (linear)
+        self.Da = np.diag([ba,ba,ba])
+        
+        
+    ##############################
+    def Tlosses(self, dq = np.zeros(3) , ddq = np.zeros(3)):
+        """ Computed torques losses given a trajectory  """  
+                
+        T = np.dot( self.Ia  , ddq ) + np.dot( self.Da  , dq )
+        
+        return T
+        
+        
+    ##############################
+    def T(self, q = np.zeros(3) , dq = np.zeros(3) , ddq = np.zeros(3) , uD = 0 ):
+        """ Computed acutator torques given a trajectory and gear ratio """ 
+        
+        R = self.R[uD]
+        
+        F = self.F( q , dq , ddq )
+        
+        Tl = self.Tlosses( dq , ddq )
+        
+        T = np.dot( np.linalg.inv(R) , F ) + np.dot( R , Tl ) 
+        
+        return T
+        
+        
+    ##############################
+    def ddq_a(self, q = np.zeros(3) , dq = np.zeros(3) , T = np.zeros(3) , uD = 0 ):
+        """ Computed accelerations given actuator torques and gear ratio """  
+        
+        R = self.R[ int(uD) ]
+        
+        Ha = self.H( q ) + np.dot( R , np.dot( R , self.Ia ) )
+        
+        C  = self.C( q , dq )
+        D  = self.D( q , dq )
+        
+        Ca =  C + D + np.dot( R , np.dot( R , self.Da ) )
+        
+        
+        G  = self.G( q )
+        
+        ddq = np.dot( np.linalg.inv( Ha ) ,  ( np.dot( R , T ) - np.dot( Ca , dq ) - G ) )
+        
+        return ddq
+        
+        
+    #############################
+    def fc(self, x = np.zeros(6) , u = np.zeros(4) , t = 0 ):
+        """ 
+        Continuous time function evaluation
+        
+        INPUTS
+        x  : state vector             n x 1
+        u  : control inputs vector    m x 1
+        t  : time                     1 x 1
+        
+        OUPUTS
+        dx : state derivative vectror n x 1
+        
+        """
+        
+        dx = np.zeros(self.n) # State derivative vector
+        
+        [ q , dq ] = self.x2q( x )   # from state vector (x) to angle and speeds (q,dq)
+        
+        ddq = self.ddq_a( q , dq , u[0:self.dof] , u[self.dof] )
+        
+        dx = self.q2x( dq , ddq )    # from angle and speeds diff (dq,ddq) to state vector diff (dx)
+        
+        return dx
+        
+        
+    ##############################
+    def e_kinetic(self, q = np.zeros(3) , dq = np.zeros(3) , uD = 0 ):
+        """ Compute kinetic energy of manipulator """  
+        
+        R = self.R[ int(uD) ]
+        
+        Ha = self.H( q ) + np.dot( R , np.dot( R , self.Ia ) )        
+        
+        e_k = 0.5 * np.dot( dq , np.dot( Ha , dq ) )
+        
+        return e_k
+        
+        
+        
+        
+'''
+#################################################################
+##################          Main                         ########
+#################################################################
+'''
+
+
+if __name__ == "__main__":     
+    """ MAIN TEST """
+    
+    R1 = HybridOneLinkManipulator()
+    R2 = HybridTwoLinkManipulator()
+    R3 = HybridThreeLinkManipulator()
