+++ projects/ICRA2017/rmin.py
