+++ projects/dLabRobots/BoeingArm.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Aug 12 10:10:34 2016
+
+@author: agirard
+"""
+
+from AlexRobotics.dynamic  import Manipulator   as M
+
+import numpy as np
+import matplotlib.pyplot as plt
