+++ projects/hybrid-robots/Output/x_B.png
